test: extract test's fixtures
